refactor: Cookie sid를 부여하는 부분을 login handler에서 처리되도록 구현 - 기존 Register 을 하면, Sid를 설정해놓는 것 수정
