+++ src/main/java/webserver/handler/response/query/RegisterHandler.java
@@ -21,9 +21,7 @@ public class RegisterHandler {
     private static final String SPACE = " ";
     private static final String CRLF = "\r\n";
 
-    private static final int SID_LENGTH = 20;
-    private static final int ASCII_START = 60;
-    private static final int ASCII_LENGTH = 66;
+
     private String requestBody;
 
     public RegisterHandler(String requestBody) {
@@ -34,8 +32,6 @@ public class RegisterHandler {
         User user = createUser();   // User 객체를 생성
         Database.addUser(user); // DB에 저장
         logger.debug("user created : {}", user);
-        String randomSid = getRandomSid();
-        SessionStore.addSession(randomSid, user);  // Session 저장소에 저장
     }
 
     public HttpResponse getResponse() {
@@ -60,20 +56,5 @@ public class RegisterHandler {
     }
 
 
-    private String getRandomSid() {
-        while (true) {
-            StringBuilder sb = new StringBuilder(SID_LENGTH);
-            Random random = new Random();
-
-            for (int i = 0; i < SID_LENGTH; i++) {
-                char randomChar = (char) (ASCII_START + random.nextInt(ASCII_LENGTH));  // 32 ~ 126 까지의 ascii code
-                sb.append(randomChar);
-            }
-
-            if (!SessionStore.sessionIdExists(sb.toString())) {
-                return sb.toString();
-            }
-        }
-    }
 
 }
